First 50 terms of Fibonacci sequence

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,15 +7,15 @@
  */
 int main(void)
 {
-	int x1 = 1, x2 = 2, x3;
+	long int x1 = 1, x2 = 2, x3;
 	int i = 0;
 
-	printf("%d, %d", x1, x2);
+	printf("%ld, %ld", x1, x2);
 	while (i < 48)
 	{
 		x3 = x1 + x2;
 		i++;
-		printf(", %d", x3);
+		printf(", %ld", x3);
 		x1 = x2;
 		x2 = x3;
 	}
